Update README and rm references to original role in comments
Includes some basic helpful information about the role in the readme.

Also removes the references to the original role name that were included
in comments to avoid confusion about using the forked version. Also
removes the galaxy meta info related to the original role.

(Would be stoked to get these changes in to the original role. I have a
PR open there and we'll see if the original author is open to accepting
some of these changes)